requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pymysql
cryptography
sentence-transformers
transformers
elasticsearch
python-dotenv
torch
tqdm
